MIPS: Loongson: CS5536: Cleanup the vendor id information
This patch removes the duplicated macros from cs5536_pci.h and merges
the switch entries for the vendor id information and the subsystem
vendor id information.

Reference:

"A company requires only one Vendor ID. That value can be used in either the
Vendor ID field of configuration space (offset 00h) or the Subsystem Vendor
ID field of configuration space (offset 2Ch). It is used in the Vendor ID
field (offset 00h) if the company built the silicon. It is used in the
Subsystem Vendor ID field (offset 2Ch) if the company built the add-in card.
If a company builds both the silicon and the add-in card, then the same
value would be used in both fields."

-- http://www.dinigroup.com/pipermail/pci-sig/2008-February/000228.html

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>