MIPS: Loongson: Enable sm501-pwm based backlight driver for gdium
CONFIG_BACKLIGHT_PWM is necessary for pwm based backlight driver, so,
add it. By the way, for the gdium hid driver is added as
module(CONFIG_HID_GDIUM=m) by default, so, no need to add it explicitly,
and also for the other hid drivers are useless in our platform, removes
them.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>