MIPS: scall64-o32.S: Emulate rdtsc instruction through "syscall code"
Lots of applications need fast rdtsc to get the high resolution/precise
time. e.g. This may be used for RNG proposed in the paper: "Analysis of
inherent randomness of the Linux kernel". and also, lots of user-space
applications can be speeded up with this rdtsc-like support.

*Example*

With this patch, users can access the count register in the user-space:

	#include <stdio.h>
	#include <stdint.h>

	/*
	 * Currently, our return value is only 32bit, In the long run,
	 * this should be uint64_t, just like clock_gettime(), but it
	 * should has high precision/low overhead than clock_gettime()
	 */
	uint32_t rdtsc(void)
	{
		__asm__ __volatile__ (
			"syscall 1\n"
		:::"$2");
	}

	int main(int argc, char *argv[])
	{
		return printf("cycles: %lu\n", rdtsc());
	}

*TODO*

1. Handle the 32bit over-flow problem in kernel-space or user-space, for
64bit count(e.g. Cavium), no need to cope with the 32bit over-flow
problem.

2. More light-weight and frequently-used operations can be abstracted
and added as fast system call. This may improve the real-time
performance a little.

3. Add rdtsc for MIPS scall64-{n32,64}.S, scall32-32.S

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>