MIPS: Loongson: Gdium: Fix the sm501pwm based clock
This patch make the sm501pwm basic clock itself work but this also can
not work as the system clocksource, which is explained in the header
file of arch/mips/loongson/gdium/gdium-clock.c. But this may be a good
clock source for some applications, such as testing the interrupt
latency of the real time system ... so, keep it work as an using
exammple. If you need it, please remove the dependency of the BROKEN in
arch/mips/loongson/Kconfig.

Since this externl clock doesn't work, so, to ensure loongson2_cpufreq
work, we must the R4K Timer work with dynamic cpu frequency, which
means, we must ensure it work normally even the hardware clock source
changes all the time. Arnaud Patard <apatard@mandriva.com> has written a
patch(Fix r4k clocksource and clockevent with cpufreq) to fix it but
that patch doesn't work well for the time still be broken, therefore, we
must rewrite it our own, we maybe possible to apply the method used in
the generic-trace-clock of Lttng contributed by Mathieu Desnoyers
<mathieu.desnoyers@efficios.com>.

The Loongson2 cpufreq driver must be enabled before the other drivers
for Gdium has overclocking the frequency to ~1G, if no cpufreq driver
support, it will be very hot and may break the machine and at least
consume more power or make the fan rotate very fast and becomes noisy.
This should be an urgent task.

The 2nd urgent task is add the Hibernation/Suspend support, which may
also save the power. The last task is add the Audio/Sound support.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>