MIPS: Loongson: Make R4K Timer Work with CPUFreq Driver
The frequency of the R4K Timer on Loongson is half of the processors'
frequency and it changes accordlingly to the processors's frequency.

If the CPUFreq driver is enabled, the R4K Timer will be 'mussy' and
therefore can not be used as the system clocksource. So, before, the
external RTC devices are added.

The overclocking of Loongson2F's frequency to 891M has made the netbook
too hot, The CPUFreq driver must be enabled to make it stay in a lower
frequency most of time, but unfortuanately, there is no suitable
external RTC driver(sm501-pwm based clock is initialized too later and
the RTC driver is similar and also both of them has lower precision and
doesn't support one-shot mode and will make CPUFreq driver doesn't work
normally) on Gdium, so, the R4K Timer becomes the last choice.

Since the R4K Timer's frequency is relevant to the processor's
frequency, If we can track the change of the processor's frequency and
record the time in every frequency and convert them to the the real
time, we may make R4K Timer work, similarly, we also need to convert the
delta of the next event to the time at the traget frequency to ensure a
task can be scheduled at the expected time.

The whole problem is splitted to two problems, the clocksource and
clockevent.

To solve the clocksource problem, a virtual 64bit count is added, and
the old c0_hpt_read() is therefore converted to a 64bit version. 64bit
here is used to avoid the 'quick' overflow of the old 32bit version.

Before, we simply read the current c0 count register, now, we record the
time we read every time and save it into a static/global variable, name
it as hpt_last_read, and therefore, we can easily record the time elaspe
between two read, If we accumulate all of them, we can get the whole
time line(hpt_last_cnt, the return value). But for the Timer's frequency
is changed all the time, so, simply accumulate them is not enough, we
must convert the timeslice(diff) to the real one. To archieve this goal,
we must know the frequency when we read the current time, this is very
easy to be finished by the CPUFreq notifiers. Similarly, the delta to
the next event can be converted to the real moment.

But the timeslice we read may cover several different frequency, so, we
must ensure every timeslice is binded with the same frequency to make
the convertion be possible. To archieve this, the new c0_hpt_read() must
be called *just* before the CPUFreq want to change the Frequency and
similarly, the delta must be updated *just* after the CPUFreq want to
change the Frequency. both of them can be finished by the CPUFreq
notifier.

For this specific implementation, some improvementation/bugfix is
introduced:

1. To simplify the conversion, we use the scale_shift(right or left shift
bits) and therefore need to choose the suitable Frequencies to ensure
the max frequency is the integral multiple of them and also make the
multiple can be converted to simple left and right shift.

The current struct cpufreq_freqs doesn't provide the related argument to
pass the scale_shift value, we simply use the global variable & function
instead of the notifier. as a result, with the irq and preemption
disabled, the related CPUFreq changes can be tracked timely.

2. The old cpu_wait() may put the cpu into wait mode and therefore may
stop the R4K Timer, so, this timeslice can not be recorded by the R4K
Timer itself and therefore external Timer is needed, but as we know, the
external timers's precision is bad, so, we must disable the cpu_wait()
for this feature. and similarly, when the system is in
suspend/hiberation status, the R4K Timer will also be stopped, so, we
must record the timelice, for if the time stay in suspend/hiberation is
theoretically longer, the precision of the external timer may not
influence too big, this will be covered by the timekeeping_resume() with
the help of read_persistent_clock(), and therefore, the virtual count
will not be 'continuous', but who will call this directly even before
and after the suspend/hibertation?!

3. Based on this new virtual 64bit count, High Resolution sched_clock()
can be easily supported, which means the high resolution sched_clock()
works with CPUFreq now, no need to set it as an option, just enable it
by default.

*NOTE*:

The basic idea is heavily basc on the commit: "Fix r4k clocksource and
clockevent with cpufreq" from
git://git.opensde.net/users/amery/linux-2.6.29-gdium, It is contributed
by Arnaud Patard <apatard@mandriva.com>. But this patch has fixed
several important potential problems and add some new features and at
last make it really works.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>