MIPS: Loongson: Add cpu_wait() for R4K+CPUFreq
The traditional cpu_wait(Stop the CPU) doesn't work with R4K+CPUFreq, so
we don't stop the CPU but put it into the 1st level of frequency
instead.

cpu_wait() is called in system idle status, so, we keep it as light
weight as we can, and also, to make Ftrace doesn't get the useless
tracing result, we notrace all of the functions called by cpu_wait().

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>