MIPS: Loongson: Gdium: Add support for old versions
Gdium has different versions, this patch adds a new option to allow the
users to choose the support for their boards.

In the future, this option will be removed and replaced by the kernel
automatical probing support with the help of FDT or machtype.

BTW: According to git://git.opensde.net/users/amery/linux-2.6.29-gdium,
The pci irq map has been fixed.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>