MIPS: Loongson: Enalbe R4K+CPUFreq+ondemand governor
The current MFGPT clocksource implementation is broken for it may lose about 12s per
hour, but CPUFreq doesn't work well with the old MFGPT clocksource. so,
we just give up the MFGPT clocksource.

And for R4K now works with CPUFreq driver and it can provides a very
good clockcource, and it works well with the ondeman governor. so, this
becomes our best choice.

To avoid the users enable the old MFGPT Timer, just disable it but to
allow people to fix it, reserve the souce code in the tree.

NOTE: CPU_FREQ_GOV_USERSPACE will not work with the latest CPUFreq
driver for it has enabled the lowest frequency as the cpu_wait() mode,
then, whenever you set the frequency, the frequency may be changed to
the lowest when the cpu becomes idle.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>