MIPS: PCI: Eliminate the warning about 'const struct pci_dev'
'const struct pci_dev' is used to forbit the modification of pci_dev
pointer, but the prototype of pcibios_map_irq used "struct pci_dev"
directly, which may be possible to introduce some problems. but the
current implementation should have no such problems, we can be safe to
remove the const. As a result, the warning about un-incompatible
arguments is eliminated.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>