vt: Make VT_BUF_VRAM_ONLY be configurable
If only need the X session, we can disable the output to the console,
then, the framebuffer driver(CONFIG_FB) can be disabled. If framebuffer
driver is disabled, VT_BUF_VRAM_ONLY can be enabled even if it doesn't
work for we don't care about the console output when CONFIG_FB=n.

But If you need normal console output, VT_BUF_VRAM_ONLY may need to be
disabled for it has been observed not work on YeeLoong laptop.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>