scripts: add undef_func.sh to print the undefined functions
This tool prints the undefined functions in order of called times, to
get the position of those functions, please use ctags or cscope.

This tool is written for the combined compiling support.

$ make ARCH=mips CROSS_COMPILE=mipsel-unknown-linux-uclibc- -j2 2>typescript
$ scripts/undef_func.sh

After get the undefined functions, please get their positions with the
help of ctags or cscope and at last add __global to mark the functions
as externally_visible.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>