Add __global tag where needed.
From: http://www.gossamer-threads.com/lists/linux/kernel/680202?#680202

This patch just adds the __global tag to functions and variables which are used
from outside the directory in which they're provided, to prevent
-fwhole-program from eating them.

If we just define __global as a dummy for now, then there's no harm in applying
this patch as-is -- but it's largely pointless without the Makefile changes.

TODO: We need to add more for the latest kernel and for MIPS/Loongson.

Note: The 4th patch will not be applied, but which may be a good
reference for future developement, it is:
http://www.gossamer-threads.com/lists/linux/kernel/680206?#680206

Signed-off-by: David Woodhouse <dwmw2@infradead.org>
[ Please get more undefined functions with the help of
scripts/undef_func.sh ]
Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>