MIPS: Kexec: Enhance the support
Changes:
  o Print more information in machine_kexec() for debugging
    E.g. with this information, the kexec_start_address has been found
    it was wrong with 64bit kernel / o32 kexec-tools. Which must be
    fixed later.
  o Link relocate_kernel.S to a section for future extension
    This allows more functions can be added for the kexec relocation
    part even written in C. to add code into that section, you just need
    to mark your function or data with __kexec or
    __attribute__((__section__(".__kexec.relocate")))

TODO:

1. Make 64bit kernel / o32|n32|64 kexec-tools works

Fix the user-space kexec-tools, seems the tool only work for 32bit
machine. So, we need to add 64bit support for it. The address of the
entry point(kexec_start_address) is wrong and make the "kexec -e" fail.
the real entry point must be read from the new kernel image by the
user-space kexec-tools, otherwise, it will not work.  The above 64bit
support tested is 64bit kernel with o32 user-space kexec-tools. The root
cause may be the different definition of virt_to_phys() and
phys_to_virt() in the kexec-tools and kernel space for 64bit system /
o32 kernel.

Ref: http://www.linux-mips.org/archives/linux-mips/2009-08/msg00149.html

2. Pass the arguments from kexec-tools to the new kernel image

Please refer to: "MIPS: Loongson: Kexec: Pass parameters to new kernel"

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>