MIPS: Loongson: Clean up the env & arg handling for EMBEDDED
This removes the loongson_cmdline[] array and use an ec_version variable
instead to reduce some size and also this allows users override the
cmdline with CONFIG_CMDLINE_OVERRIDE, then can remove the whole arg
probing codes.

Similarly, the env vars can be overrided with the help of
CONFIG_ENVVAR_OVERRIDE. when CONFIG_ENVVAR_OVERRIDE is enable, the
memory size and cpu freq can be configurable.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>