tracing/MIPS: Fix set_graph_function of function graph tracer
trace.func should be set as the 'selfpc'(The recorded ip in the
__mcount_loc section) to match the input of set_graph_function, but
before, this is set as the self_ra(the return address of mcount), which
has made set_graph_function not work as expected.

This fixes it via calculating the right recorded ip in the __mcount_loc
section and assign it to trace.func.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>