MIPS: Flush cache before calling the decompressed kernel image
A 'crude' cache flush operation is added for compressed kernel image
support to make sure the to be run code in cache are what we expected.

Without this patch, when with small vmlinux, vmlinuz will fail in most
of time, but for big vmlinux, it success in most of the time.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>