MIPS: Loongson: add nano_defconfig for YeeLoong
This nano_defconfig only support console & kexec, which may allow it
work as a small bootloader.

Note: kexec needs the /proc/iomem interface to get the memory ranges,
but to save the size we have disabled procfs, so, we need to save the
content of /proc/iomem to another file and let kexec use it instead, for
example:

$ cat /proc/iomem > /iomem.txt

And then, change the /proc/iomem in kexec/arch/mips/kexec-mips.c to
/iomem.txt. but this is not flexible, perhaps we'd better provide
another device under /dev/iomem to fix this problme later.

Now, the kernel size is:

$ ls -lh vmlinuz
-rwxr-xr-x 1 root root 418K 2011-01-21 03:34 vmlinuz

And we can boot another kernel with this kernel loaded.

$ git clone git://git.kernel.org/pub/scm/utils/kernel/kexec/kexec-tools.git
$ cd kexec-tools

Mark virt_to_phys() function defined in kexec/virt_to_phys.c as week

-unsigned long virt_to_phys(unsigned long UNUSED(addr))
+unsigned long __attribute((weak)) virt_to_phys(unsigned long UNUSED(addr))

Then, follow the above steps to prepare a /iomem.txt and change the
/proc/iomem in kexec/arch/mips/kexec-mips.c to /iomem.txt

Then, compile it

$ make

Load a new kernel image

$ build/sbin/kexec -l ./vmlinux

Execute it

$ build/sbin/kexec -e

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>