param: Make built-in module parameters configurable
Like the generic kernel parameters, the module parameters may also never
be used, this make them be configurable. the grain is all of the
modules, not the specific parameter. The specific parameter grain will
be the future work, which can be easily implemented with the reference
to the generic parameters.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>