scripts: add sizereport.sh
This adds a shell script to report the size distributing of the kernel
image. This report is not enough, more details will be reported in the
enhancement of this shell script, for example, in the specific object,
the biggest function should be reported. A long-term of this shell
script should become a system, which can be linked together with the xlr
online system or the cscope or ctags system(with the help of the vim
shortcuts).

This is very useful for future size reducing of the kernel image size,
As oprofile function as a profiling tool, this may function as a size
reducing tool, but it is far from global function ;-)

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>