MIPS: Lemote: Enable rtl8187 in defconfig
rtl8187 driver is used instead of the old rtl8187b driver in the default
defconfig, and correspondingly, the LED support is enabled.

Note: rtl8187b is almost broken and it only work with WPA & WPA personal
currently and I have no plan to fix the existing problems. so, please
use rtl8187 instead if you are not using WPA & WPA personal.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>